+++ spring-cloud-starter-tencent-metadata-transfer/src/main/java/com/tencent/cloud/metadata/provider/RestTemplateMetadataProvider.java
@@ -40,7 +40,7 @@ public class RestTemplateMetadataProvider implements MetadataProvider {
 	}
 
 	@Override
-	public String getRawMetadataStringValue(String key) {
+	public String doGetRawMetadataStringValue(String key) {
 		switch (key) {
 		case MessageMetadataContainer.LABEL_KEY_METHOD:
 			return request.getMethod().toString();
@@ -55,7 +55,7 @@ public class RestTemplateMetadataProvider implements MetadataProvider {
 	}
 
 	@Override
-	public String getRawMetadataMapValue(String key, String mapKey) {
+	public String doGetRawMetadataMapValue(String key, String mapKey) {
 		switch (key) {
 		case MessageMetadataContainer.LABEL_MAP_KEY_HEADER:
 			return UrlUtils.decode(SpringWebExpressionLabelUtils.getHeaderValue(request, mapKey));