+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - [feat: upgrade springdoc to 1.8.0.](https://github.com/Tencent/spring-cloud-tencent/pull/1737)
 - [refactor:optimize auto configuration.](https://github.com/Tencent/spring-cloud-tencent/pull/1740)
 - [refactor:optimize config locate.](https://github.com/Tencent/spring-cloud-tencent/pull/1742)
+- [feat:support async metadata transfer.](https://github.com/Tencent/spring-cloud-tencent/pull/1744)